修改frame，frameHandler,frameDispatcher接口定义，

+++ protocol/src/main/java/org/archer/archermq/protocol/constants/FrameTypeEnum.java
@@ -7,27 +7,29 @@ import java.util.Objects;
 
 /**
  * 数据帧类型枚举
- * @date 2020年04月13日22:41:14
+ *
  * @author dongyue
+ * @date 2020年04月13日22:41:14
  */
-public enum FrameTypeEnum implements EnumSpec<Integer> {
+public enum FrameTypeEnum implements EnumSpec<Byte> {
 
-    METHOD(1,"method frame"),
-    CONTENT_HEADER(2,"content header frame"),
-    CONTENT_BODY(3,"content body frame"),
-    HEARTBEAT(8,"heart beat frame");
+    METHOD((byte) 1, "method frame"),
+    CONTENT_HEADER((byte) 2, "content header frame"),
+    CONTENT_BODY((byte) 3, "content body frame"),
+    HEARTBEAT((byte) 8, "heart beat frame");
 
 
-    private final int val;
+    private final Byte val;
     private final String desc;
 
-    FrameTypeEnum(int val, String desc) {
+    FrameTypeEnum(byte val, String desc) {
         this.val = val;
         this.desc = desc;
     }
 
+
     @Override
-    public Integer getVal() {
+    public Byte getVal() {
         return val;
     }
 
@@ -36,10 +38,10 @@ public enum FrameTypeEnum implements EnumSpec<Integer> {
         return desc;
     }
 
-    public static FrameTypeEnum getByVal(Integer val){
+    public static FrameTypeEnum getByVal(Byte val) {
         FrameTypeEnum[] types = values();
-        for(FrameTypeEnum type:types){
-            if(Objects.equals(val,type.val)){
+        for (FrameTypeEnum type : types) {
+            if (Objects.equals(val, type.val)) {
                 return type;
             }
         }
